backend/requirements.txt
# Core Flask and API dependencies
Flask==3.1.0
Flask-CORS==5.0.0
python-dotenv==1.1.1

# Reddit API
praw==7.8.1
prawcore==2.4.0

# Machine Learning and NLP
transformers==4.53.2
torch==2.7.1
scikit-learn==1.6.1
numpy==2.2.4
pandas==2.2.3

# HTTP requests
requests==2.32.3

# Text processing
nltk==3.9.1
regex==2024.11.6

# Utilities and dependencies
packaging==24.2
tqdm==4.67.1
huggingface-hub==0.33.4
tokenizers==0.21.2
safetensors==0.5.3
filelock==3.18.0
click==8.2.1
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
